+++ packages/contracts/contracts/Dependencies/BaseMath.sol
@@ -3,5 +3,5 @@ pragma solidity ^0.8.17;
 
 
 contract BaseMath {
-    uint constant public DECIMAL_PRECISION = 1e18;
+    uint256 constant public DECIMAL_PRECISION = 1e18;
 }
